move common customisation functions to a shared place
